the original version of combat scene
in the combat scene, the different areas are used to display different objcets.This version could draw the backgroud of the main several areas like role Information,cost,role card and tools and all of these are achieved by functions. When the UI group try to do their work, these functions could be used and modified.